Refactor: Organized model interfaces/classes
